Split orchestration protocols from DTOs
Move orchestration protocol definitions and usage aggregation helpers out
of `_dto.py` so the DTO module stays under the file-size limit and keeps a
single responsibility.

Consolidate duplicate enum-rejection tests across planned actions and
action execution results into one parametrized validation path.
