+++ episodic/orchestration/_dto.py
@@ -396,71 +396,3 @@ class GenerationOrchestrationResult:
                 )
                 raise TypeError(msg)
         object.__setattr__(self, "action_results", action_results)
-
-
-class ToolExecutorPort(typ.Protocol):
-    """Application-level port for executing planned enrichment actions."""
-
-    async def execute(
-        self,
-        action: PlannedAction,
-        context: GenerationOrchestrationRequest,
-    ) -> ActionExecutionResult:
-        """Execute one planned action against the available generation context."""
-
-
-class PlannerPort(typ.Protocol):
-    """Application-level port for structured orchestration planning."""
-
-    async def plan(
-        self,
-        request: GenerationOrchestrationRequest,
-    ) -> PlannerResult:
-        """Return a typed execution plan for the supplied generation request."""
-
-
-class _ShowNotesGeneratorPort(typ.Protocol):
-    """Abstraction for show-notes generation used by the first tool executor."""
-
-    async def generate(
-        self,
-        script_tei_xml: str,
-        *,
-        template_structure: dict[str, object] | None = None,
-    ) -> ShowNotesResult:
-        """Generate show notes from the supplied TEI context."""
-
-
-def _sum_usage(*usage_values: LLMUsage | None) -> LLMUsage:
-    """Return the total token usage across all provided LLMUsage records."""
-    input_tokens = 0
-    output_tokens = 0
-    total_tokens = 0
-    for usage in usage_values:
-        if usage is None:
-            continue
-        input_tokens += usage.input_tokens
-        output_tokens += usage.output_tokens
-        total_tokens += usage.total_tokens
-    return LLMUsage(
-        input_tokens=input_tokens,
-        output_tokens=output_tokens,
-        total_tokens=total_tokens,
-    )
-
-
-def build_generation_result(
-    planner_result: PlannerResult,
-    action_results: tuple[ActionExecutionResult, ...],
-) -> GenerationOrchestrationResult:
-    """Aggregate planner and tool usage into one orchestration result."""
-    total_usage = _sum_usage(
-        planner_result.usage,
-        *(action_result.usage for action_result in action_results),
-    )
-    return GenerationOrchestrationResult(
-        plan=planner_result.plan,
-        action_results=action_results,
-        planner_usage=planner_result.usage,
-        total_usage=total_usage,
-    )
